chore: Zcl cluster definition, extract ClusterDefinition interface
