+++ src/zcl/definition/cluster.ts
@@ -1,32 +1,8 @@
 /* eslint max-len: 0 */
 import DataType from './dataType';
-import {ParameterDefinition} from './tstype';
+import {ClusterDefinition} from './tstype';
 import BuffaloZclDataType from './buffaloZclDataType';
 import ManufacturerCode from './manufacturerCode';
-
-interface AttributeDefinition {
-    ID: number;
-    type: DataType;
-    manufacturerCode?: number;
-}
-
-interface ClusterDefinition {
-    ID: number;
-    manufacturerCode?: number;
-    attributes: {[s: string]: AttributeDefinition};
-    commands: {
-        [s: string]: CommandDefinition;
-    };
-    commandsResponse: {
-        [s: string]: CommandDefinition;
-    };
-}
-
-interface CommandDefinition {
-    ID: number;
-    parameters: ParameterDefinition[];
-    response?: number;
-}
 
 const Cluster: {
     [s: string]: ClusterDefinition;
